updated the routes so the admin can update articles, stash them, edit stash separately and publish the edited work

+++ routes/admin/article.js
@@ -12,8 +12,15 @@ router.post('/api/admin/article',async (req,res)=>{
     const {body}=req;
     if(body.stash_id){
       const id=new Types.ObjectId(`${body.stash_id}`);
+      const updatedStash=await Stash.findByIdAndUpdate(id,{...body,stash_id:undefined},{runValidators:true});
+      if(!updatedStash) throw new Error('Error while updating the stashed work');
       const deletedStash= await Stash.findByIdAndDelete(id);
       if(!deletedStash) throw new Error('Error while deleting the stashed work');
+      if(deletedStash.from_article){
+        const updatedArticle=await Article.findByIdAndUpdate(deletedStash.from_article,{title:deletedStash.title,summary:deletedStash.summary,content:deletedStash.content},{runValidators:true});
+        if(!updatedArticle) throw new Error('Error while updating an already existing article from stash');
+        return res.status(200).send({success:true});
+      }
     }
     const newArticle=new Article({...body,stash_id:undefined});
     const savedArticle= await newArticle.save();
@@ -22,6 +29,30 @@ router.post('/api/admin/article',async (req,res)=>{
   }catch(err){
     console.log(err);
     return res.status(400).send({message:err})
+  }
+});
+
+router.post('/api/admin/article/update',async (req,res)=>{
+  const {body}=req;
+  const id=new Types.ObjectId(`${body.article_id}`);
+  try{
+    const updatedArticle=await Article.findByIdAndUpdate(id,{...body,article_id:undefined},{runValidators:true});
+    if(!updatedArticle) throw new Error('Error when updating an article');
+    return res.status(200).send({success:true});
+  }catch(err){
+    console.log(err);
+    res.status(400).send({error:err});
+  }
+});
+router.get('/api/admin/articles/:id',async(req,res)=>{
+  const id=new Types.ObjectId(`${req.params.id}`);
+  try{
+    const article=await Article.findById(id,{content:1,title:1,summary:1,_id:1});
+    if(!article) throw new Error('Error occured when fetching the article data');
+    res.status(200).send({article});
+  }catch(err){
+    console.log(err);
+    res.status(400).send({error:err});
   }
 });
 export default router
